add password restore handler
