+++ cmd/backend/server/handlers/user_verify.go
@@ -2,9 +2,7 @@ package handlers
 
 import (
 	"backend/internal/core/services"
-	httpserver "backend/internal/http_server"
 	"backend/pkg/logger"
-	"context"
 
 	"html/template"
 
@@ -70,17 +68,4 @@ func NewUserVerifyEmailHandler(log logger.Logger, userService services.UserServi
 		template.Execute(c.Writer, tmp)
 		c.Status(200)
 	}
-}
-
-type inputSendVerify struct {
-	Email string `json:"email" validate:"required,email"`
-}
-
-func NewUserSendVerifyEmailHandler(log logger.Logger, userService services.UserService) gin.HandlerFunc {
-	return httpserver.WrapGin(log,
-		func(ctx context.Context, input inputSendVerify) (interface{}, error) {
-			err := userService.SendEmailVerifyEmail(ctx, input.Email)
-			return nil, err
-		},
-	)
 }
